BSeries.jl API

BSeries.ExactSolution — Type

ExactSolution{V}()

Lazy representation of the B-series of the exact solution of an ordinary differential equation using coefficients of type at least as representative as V.

BSeries.ExactSolution — Method

ExactSolution(series_integrator)

A representation of the B-series of the exact solution of an ODE using the same type of coefficients as the B-series series_integrator.

BSeries.TruncatedBSeries — Type

TruncatedBSeries

An struct that can describe B-series of both numerical integration methods (where the coefficient of the empty tree is unity) and right-hand sides of an ordinary differential equation and perturbations thereof (where the coefficient of the empty tree is zero) up to a prescribed order.

Generally, this kind of struct should be constructed via bseries or one of the other functions returning a B-series, e.g., modified_equation or modifying_integrator.

BSeries.bseries — Method

bseries(A::AbstractMatrix, b::AbstractVector, c::AbstractVector, order)

Compute the B-series of the Runge-Kutta method with Butcher coefficients A, b, c up to a prescribed integer order.

Normalization by elementary differentials

The coefficients of the B-series returned by this method need to be multiplied by a power of the time step divided by the symmetry of the rooted tree and multiplied by the corresponding elementary differential of the input vector field $f$. See also evaluate.

BSeries.compose — Method

compose(b, a, t::RootedTree)

Compute the coefficient correspoding to the tree t of the B-series that is formed by composing the B-series a with the B-series b. It is assumed that the B-series b has the coefficient unity of the empty tree.

References

Section 3.1 of

Philippe Chartier, Ernst Hairer, Gilles Vilmart (2010) Algebraic Structures of B-series. Foundations of Computational Mathematics DOI: 10.1007/s10208-010-9065-1

BSeries.compose — Method

compose(b, a; normalize_stepsize=false)

Compose the B-series a with the B-series b. It is assumed that the B-series b has the coefficient unity of the empty tree.

If normalize_stepsize = true, the coefficients of the returned B-series will are divied by 2^order(t) for each rooted tree t. This normalizes the step size so that the resulting numerical integrator B-series uses the same step size as the input series.

References

Section 3.1 of

Philippe Chartier, Ernst Hairer, Gilles Vilmart (2010) Algebraic Structures of B-series. Foundations of Computational Mathematics DOI: 10.1007/s10208-010-9065-1

BSeries.compute_derivative — Function

compute_derivative(expression, variable)

Internal function specialized on symbolic variables and expressions from

SymEngine.jl,

SymPy.jl, and

Symbolics.jl

if these packages are loaded (via Requires.jl).

BSeries.elementary_differentials — Method

elementary_differentials(f, u, order)

Compute all elementary differentials of the vector field f with independent variables u up to the given order. The return value can be indexed by rooted trees to obtain the corresponding elementary differential.

BSeries.evaluate — Function

evaluate(f, u, dt, series, reduce_order_by=0)

Evaluate the B-series series specialized to the ordinary differential equation $u'(t) = f(u(t))$ with vector field f and dependent variables u for a time step size dt.

Here, u is assumed to be a vector of symbolic variables and f is assumed to be a vector of expressions in these variables. Currently, symbolic variables from

SymEngine.jl,

SymPy.jl, and

Symbolics.jl

are supported.

The powers of dt can be controled by reduce_order_by to make them different from the usual order(t) for a rooted tree t. This can be useful in the context of modified_equations or modifying_integrators, where the B-series coeffcients are those of $h fₕ$, i.e., they contain an additional power of dt. In this case, the B-series of the vector field can be obtained using reduce_order_by = 1.

References

Section 3.2 of

Philippe Chartier, Ernst Hairer, Gilles Vilmart (2010) Algebraic Structures of B-series. Foundations of Computational Mathematics DOI: 10.1007/s10208-010-9065-1

BSeries.modified_equation — Method

modified_equation(A::AbstractMatrix, b::AbstractVector, c::AbstractVector,
                  order)

Compute the B-series of the modified_equation of the Runge-Kutta method with Butcher coefficients A, b, c up to the prescribed order.

Normalization by elementary differentials

The coefficients of the B-series returned by this method need to be multiplied by a power of the time step divided by the symmetry of the rooted tree and multiplied by the corresponding elementary differential of the input vector field $f$. See also evaluate.

BSeries.modified_equation — Method

modified_equation(f, u, dt,
                  A::AbstractMatrix, b::AbstractVector, c::AbstractVector,
                  order)

Compute the B-series of the modified_equation of the Runge-Kutta method with Butcher coefficients A, b, c up to the prescribed order with respect to the ordinary differential equation $u'(t) = f(u(t))$ with vector field f and dependent variables u for a time step size dt.

Here, u is assumed to be a vector of symbolic variables and f is assumed to be a vector of expressions in these variables. Currently, symbolic variables from

SymEngine.jl,

SymPy.jl, and

Symbolics.jl

are supported.

BSeries.modified_equation — Method

modified_equation(series_integrator)

Compute the B-series of the modified equation of the time integration method with B-series series_integrator.

Given an ordinary differential equation (ODE) $u'(t) = f(u(t))$ and a Runge-Kutta method, the idea is to interpret the numerical solution with given time step size as exact solution of a modified ODE $u'(t) = fₕ(u(t))$. This method returns the B-series of $h fₕ$.

Normalization by elementary differentials

The coefficients of the B-series returned by this method need to be multiplied by a power of the time step divided by the symmetry of the rooted tree and multiplied by the corresponding elementary differential of the input vector field $f$. See also evaluate.

References

Section 3.2 of

Philippe Chartier, Ernst Hairer, Gilles Vilmart (2010) Algebraic Structures of B-series. Foundations of Computational Mathematics DOI: 10.1007/s10208-010-9065-1

BSeries.modifying_integrator — Method

modifying_integrator(A::AbstractMatrix, b::AbstractVector, c::AbstractVector,
                     order)

Compute the B-series of the modifying_integrator equation of the Runge-Kutta method with Butcher coefficients A, b, c up to the prescribed order.

Normalization by elementary differentials

The coefficients of the B-series returned by this method need to be multiplied by a power of the time step divided by the symmetry of the rooted tree and multiplied by the corresponding elementary differential of the input vector field $f$. See also evaluate.

BSeries.modifying_integrator — Method

modifying_integrator(f, u, dt,
                     A::AbstractMatrix, b::AbstractVector, c::AbstractVector,
                     order)

Compute the B-series of the modifying_integrator equation of the Runge-Kutta method with Butcher coefficients A, b, c up to the prescribed order with respect to the ordinary differential equation $u'(t) = f(u(t))$ with vector field f and dependent variables u for a time step size dt.

Here, u is assumed to be a vector of symbolic variables and f is assumed to be a vector of expressions in these variables. Currently, symbolic variables from

SymEngine.jl,

SymPy.jl, and

Symbolics.jl

are supported.

BSeries.modifying_integrator — Method

modifying_integrator(series_integrator)

Compute the B-series of a "modifying integrator" equation of the time integration method with B-series series_integrator.

Given an ordinary differential equation (ODE) $u'(t) = f(u(t))$ and a Runge-Kutta method, the idea is to find a modified ODE $u'(t) = fₕ(u(t))$ such that the numerical solution with given time step size is the exact solution of the original ODE. This method returns the B-series of $h fₕ$.

Normalization by elementary differentials

The coefficients of the B-series returned by this method need to be multiplied by a power of the time step divided by the symmetry of the rooted tree and multiplied by the corresponding elementary differential of the input vector field $f$. See also evaluate.

References

Section 3.2 of

Philippe Chartier, Ernst Hairer, Gilles Vilmart (2010) Algebraic Structures of B-series. Foundations of Computational Mathematics DOI: 10.1007/s10208-010-9065-1

BSeries.substitute — Method

substitute(b, a, t::RootedTree)

Compute the coefficient correspoding to the tree t of the B-series that is formed by substituting the B-series b into the B-series a. It is assumed that the B-series b has the coefficient zero of the empty tree.

References

Section 3.2 of

Philippe Chartier, Ernst Hairer, Gilles Vilmart (2010) Algebraic Structures of B-series. Foundations of Computational Mathematics DOI: 10.1007/s10208-010-9065-1

RootedTrees.order — Method

order(series::TruncatedBSeries)

The maximal order of a rooted tree with non-vanishing coefficient in the truncated B-series series.